// src/app/shared/models/customer.model.ts
export class Customer {
    id: number = 0;    
    cust_descripcion?: string;
    nucf_grupo?: string;
    nucf_lote?: string;
    cust_contacto?: string;
    cust_website?: string;
    cust_api_url?: string;
    cust_email?: string;
    cust_documtos?: string;
    cust_autorizacion?: string;
    cust_envios?: string;
    cust_formato?: string;
    cust_medcmtos?: string;
    cust_cirugia?: string;
    cust_direccion?: string;
    cust_lug_tipo?: string;
    cust_lug_nombre?: string;
    cust_lug_numero?: string;
    cust_lug_apto?: string;
    sect_codigo?: string;
    ciud_codigo?: string;
    cust_telefono?: string;
    cust_fax?: string;
    cust_celular?: string;
    cust_beeper?: string;
    cust_contrato?: string;
    cust_limite?: string;
    cust_ini_contrato?: string;
    cust_fin_contrato?: string;
    cust_fecha_corte?: string;
    cust_porc_descto?: string;
    pago_codigo?: string;
    cust_rnc?: string;
    ctas_numero?: string;
    grup_codigo?: string;
    cust_comentario?: string;
    cust_estatus?: string;
    cust_referencia?: string;
    naci_codigo?: string;
    cust_web_site?: string;
    cust_local?: string;
    ctas_debito?: string;
    ctas_credito?: string;
    ctas_itbis?: string;
    ctas_descto?: string;
    ctas_retencion?: string;
    cust_tipo?: string;
    cust_movil?: string;
    cust_oficina?: string;
    cust_limite_credito?: string;
    cust_descto_pago?: string;
    ctas_isr?: string;
    cust_categoria?: string;
    cust_grupo?: string;
    cust_descto?: string;
    mone_codigo?: string;
    cust_apertura?: string;
    nucf_prefijo?: string;
    dgii_tipo?: string;
    inve_grupo?: string;
    porc_itbis?: string;
    porc_retencion?: string;
    dgii_costo?: string;
    clas_codigo?: string;
    dpto_codigo?: string;
    
    created?: Date;
    updated?: Date;
    userId?: number;
    status?: boolean;
}